--- src/llamafactory/data/data_converter/alpaca_converter.py
@@ -1,84 +1,0 @@
-# Copyright 2025 the LlamaFactory team.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import os
-from abc import abstractmethod
-from dataclasses import dataclass
-from typing import TYPE_CHECKING, Any, Optional, Union
-
-from ...extras import logging
-from ..data_utils import Role
-from .base_converter import DatasetConverter
-
-if TYPE_CHECKING:
-    from ..datasets import Dataset, IterableDataset
-    from ..transformers import Seq2SeqTrainingArguments
-
-    from ...hparams import DataArguments
-    from ..mm_plugin import AudioInput, ImageInput, VideoInput
-    from ..parser import DatasetAttr
-
-    MediaType = Union[ImageInput, VideoInput, AudioInput]
-
-
-logger = logging.get_logger(__name__)
-
-@dataclass
-class AlpacaDatasetConverter(DatasetConverter):
-    def __call__(self, example: dict[str, Any]) -> dict[str, Any]:
-        prompt = []
-        if self.dataset_attr.history and isinstance(example[self.dataset_attr.history], list):
-            for old_prompt, old_response in example[self.dataset_attr.history]:
-                prompt.append({"role": Role.USER.value, "content": old_prompt})
-                prompt.append({"role": Role.ASSISTANT.value, "content": old_response})
-
-        query = []
-        if self.dataset_attr.prompt and example[self.dataset_attr.prompt]:
-            query.append(example[self.dataset_attr.prompt])
-
-        if self.dataset_attr.query and example[self.dataset_attr.query]:
-            query.append(example[self.dataset_attr.query])
-
-        prompt.append({"role": Role.USER.value, "content": "\n".join(query)})  # "prompt\nquery"
-
-        if self.dataset_attr.kto_tag and isinstance(example[self.dataset_attr.kto_tag], bool):  # kto example
-            response = [{"role": Role.ASSISTANT.value, "content": example[self.dataset_attr.response]}]
-            if example[self.dataset_attr.kto_tag]:
-                response = response + [{"role": Role.ASSISTANT.value, "content": ""}]
-            else:
-                response = [{"role": Role.ASSISTANT.value, "content": ""}] + response
-        elif (
-            self.dataset_attr.ranking
-            and isinstance(example[self.dataset_attr.chosen], str)
-            and isinstance(example[self.dataset_attr.rejected], str)
-        ):  # pairwise example
-            response = [
-                {"role": Role.ASSISTANT.value, "content": example[self.dataset_attr.chosen]},
-                {"role": Role.ASSISTANT.value, "content": example[self.dataset_attr.rejected]},
-            ]
-        elif self.dataset_attr.response and isinstance(example[self.dataset_attr.response], str):  # normal example
-            response = [{"role": Role.ASSISTANT.value, "content": example[self.dataset_attr.response]}]
-        else:  # unsupervised
-            response = []
-
-        output = {
-            "_prompt": prompt,
-            "_response": response,
-            "_system": example[self.dataset_attr.system] if self.dataset_attr.system else None,
-            "_tools": example[self.dataset_attr.tools] if self.dataset_attr.tools else "",
-            "_images": self._find_medias(example[self.dataset_attr.images]) if self.dataset_attr.images else None,
-            "_videos": self._find_medias(example[self.dataset_attr.videos]) if self.dataset_attr.videos else None,
-            "_audios": self._find_medias(example[self.dataset_attr.audios]) if self.dataset_attr.audios else None,
-        }
-        return output
