--- src/llamafactory/data/data_converter/base_converter.py
@@ -1,66 +1,0 @@
-# Copyright 2025 the LlamaFactory team.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import os
-from abc import abstractmethod
-from dataclasses import dataclass
-from typing import TYPE_CHECKING, Any, Optional, Union
-
-from ...extras import logging
-from ..data_utils import Role
-
-if TYPE_CHECKING:
-    from ..datasets import Dataset, IterableDataset
-    from ..transformers import Seq2SeqTrainingArguments
-
-    from ...hparams import DataArguments
-    from ..mm_plugin import AudioInput, ImageInput, VideoInput
-    from ..parser import DatasetAttr
-
-    MediaType = Union[ImageInput, VideoInput, AudioInput]
-
-
-logger = logging.get_logger(__name__)
-
-
-@dataclass
-class DatasetConverter:
-    dataset_attr: "DatasetAttr"
-    data_args: "DataArguments"
-
-    def _find_medias(self, medias: Union["MediaType", list["MediaType"], None]) -> Optional[list["MediaType"]]:
-        r"""Optionally concatenate media path to media dir when loading from local disk."""
-        if medias is None:
-            return None
-        elif not isinstance(medias, list):
-            medias = [medias]
-        elif len(medias) == 0:
-            return None
-        else:
-            medias = medias[:]
-
-        if self.dataset_attr.load_from in ["script", "file"] and isinstance(medias[0], str):
-            for i in range(len(medias)):
-                if os.path.isfile(os.path.join(self.data_args.media_dir, medias[i])):
-                    medias[i] = os.path.join(self.data_args.media_dir, medias[i])
-                else:
-                    logger.warning_rank0_once(f"Media {medias[i]} does not exist in `media_dir`. Use original path.")
-
-        return medias
-
-    @abstractmethod
-    def __call__(self, example: dict[str, Any]) -> dict[str, Any]:
-        r"""Convert a single example in the dataset to the standard format."""
-        ...
-
