--- src/llamafactory/data/data_converter/converter.py
@@ -1,97 +1,0 @@
-# Copyright 2025 the LlamaFactory team.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import os
-from abc import abstractmethod
-from dataclasses import dataclass
-from typing import TYPE_CHECKING, Any, Optional, Union
-
-from ...extras import logging
-from ..data_utils import Role
-from .alpaca_converter import AlpacaDatasetConverter
-from .sharegpt_converter import SharegptDatasetConverter
-from .notemplate_converter import NoTemplateDatasetConverter
-from .nlu_converter import NluHeadDatasetConverter
-
-if TYPE_CHECKING:
-    from ..datasets import Dataset, IterableDataset
-    from ..transformers import Seq2SeqTrainingArguments
-
-    from ...hparams import DataArguments
-    from ..mm_plugin import AudioInput, ImageInput, VideoInput
-    from ..parser import DatasetAttr
-
-    MediaType = Union[ImageInput, VideoInput, AudioInput]
-
-
-logger = logging.get_logger(__name__)
-
-
-DATASET_CONVERTERS = {
-    "alpaca": AlpacaDatasetConverter,
-    "sharegpt": SharegptDatasetConverter,
-    "targetnlu": NluHeadDatasetConverter,
-    "dynamic": NoTemplateDatasetConverter,
-}
-
-
-def register_dataset_converter(name: str, dataset_converter: type["DatasetConverter"]) -> None:
-    r"""Register a new dataset converter."""
-    if name in DATASET_CONVERTERS:
-        raise ValueError(f"Dataset converter {name} already exists.")
-
-    DATASET_CONVERTERS[name] = dataset_converter
-
-
-def get_dataset_converter(name: str, dataset_attr: "DatasetAttr", data_args: "DataArguments") -> "DatasetConverter":
-    r"""Get a dataset converter."""
-    if name not in DATASET_CONVERTERS:
-        raise ValueError(f"Dataset converter {name} not found.")
-
-    return DATASET_CONVERTERS[name](dataset_attr, data_args)
-
-
-def align_dataset(
-    dataset: Union["Dataset", "IterableDataset"],
-    dataset_attr: "DatasetAttr",
-    data_args: "DataArguments",
-    training_args: "Seq2SeqTrainingArguments",
-) -> Union["Dataset", "IterableDataset"]:
-    r"""Align the dataset to a specific format.
-
-    Aligned dataset:
-    _prompt: [{"role": "user", "content": "..."}] * (2T - 1)
-    _response: [{"role": "assistant", "content": "..."}] * N (N > 1 for ranking dataset)
-    _system: "..."
-    _tools: "..."
-    _images: []
-    _videos: []
-    _audios: []
-    """
-    column_names = list(next(iter(dataset)).keys())
-    kwargs = {}
-    if not data_args.streaming:
-        kwargs = dict(
-            num_proc=data_args.preprocessing_num_workers,
-            load_from_cache_file=(not data_args.overwrite_cache) or (training_args.local_process_index != 0),
-            desc="Converting format of dataset",
-        )
-
-    dataset_converter = get_dataset_converter(dataset_attr.formatting, dataset_attr, data_args)
-    return dataset.map(
-        dataset_converter,
-        batched=False,
-        remove_columns=column_names,
-        **kwargs,
-    )
